Cleaned up Makefile

+++ Makefile
@@ -2,7 +2,7 @@ SCRIPT=$(wildcard plugin/*.vim)
 FAQ="vim_faq.txt"
 DOC=$(wildcard doc/*.txt)
 PLUGIN=$(shell basename "$$PWD")
-#VERSION=$(shell sed -n '/Version:/{s/^.*\(\S\.\S\+\)$$/\1/;p}' $(SCRIPT))
+VERSION=$(shell sed -n '/Version:/{s/^.*\(\S\.\S\+\)$$/\1/;p}' $(SCRIPT))
 
 .PHONY: $(PLUGIN).vba README
 
@@ -11,7 +11,7 @@ all: generate uninstall vimball install README
 vimball: $(PLUGIN).vba
 
 clean:
-	rm -f *.vba */*.orig *.~* .VimballRecord
+	rm -f *.vba *.~* .VimballRecord
 	find . -type f -name "*.un~" -delete
 
 dist-clean: clean
@@ -22,21 +22,18 @@ install:
 uninstall:
 	vim -N -c':RmVimball' -c':q!' $(PLUGIN)-$(VERSION).vba
 
-undo:
-	for i in */*.orig; do mv -f "$$i" "$${i%.*}"; done
-
 README:
 	cp -f $(DOC) README
 
 $(PLUGIN).vba:
 	rm -f $(PLUGIN)-$(VERSION).vba
 	vim -N -c 'ru! vimballPlugin.vim' -c ':call append("0", [ "$(SCRIPT)", "$(DOC)"])' -c '$$d' -c ":%MkVimball $(PLUGIN)-$(VERSION)  ." -c':q!'
-	ln -f $(PLUGIN)-$(VERSION).vba $(PLUGIN).vba
+	if [ -f $(PLUGIN)-$(VERSION).vba ]; then ln -f $(PLUGIN)-$(VERSION).vba $(PLUGIN).vba; endif
      
 generate:
 	vim -u NONE -U NONE -N -c ':\$s/Last updated on: \zs/\=strftime("%d %B %Y")/|wq' ${FAQ}
 	vim -u NONE -U NONE -N -c ':so vim_faq.vim|:CreateVimFAQHelp|' ${FAQ}
-	#perl -i -pne 'if (/GetLatestVimScripts:/) {s/(\d+)\s+:AutoInstall:/sprintf("%d :AutoInstall:", 1+$$1)/e}' ${SCRIPT}  ${AUTOL}
-	#perl -i -pne 'if (/Last Change:/) {s/(:\s+).*\n/sprintf(": %s", `date -R`)/e}' ${SCRIPT} ${AUTOL}
-	#perl -i.orig -pne 'if (/Version:/) {s/\.(\d+).*\n/sprintf(".%d %s", 1+$$1, `date -R`)/e}' ${DOC}
-	#VERSION=$(shell sed -n '/Version:/{s/^.*\(\S\.\S\+\)$$/\1/;p}' $(SCRIPT))
+	vim -u NONE -U NONE -N -c '/^" GetLatestVimScripts: /s/\d\+\s\ze:AutoInstall:/\=submatch(0)+1/|wq' ${SCRIPT}
+	vim -u NONE -U NONE -N -c '/^" Last Change: /s/\$/strftime("%d %B %Y)/|wq' ${SCRIPT}
+	vim -u NONE -U NONE -N -c '/^" Version:/s/\d\+/\=submatch(0)+1/|wq' ${SCRIPT}
+	VERSION=$(shell sed -n '/Version:/{s/^.*\(\S\.\S\+\)$$/\1/;p}' $(SCRIPT))
